python/edtwExample/requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
filelock==3.6.0
platformdirs==2.5.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.13.1
